use 1/4th of flash for character definitions
This works because for CH32v003 you can just set BR all the time and
have BS determine the full state of a GPIO in the BSHR register.

According to 7.3.1.4 of CH32V003RM:

If both BR and BS bits are set, the BS bit takes effect.

This moves some extra work to vga_prepare_line.